third_party/WebKit/LayoutTests/animations/prefixed/animation-inherit-initial-unprefixed-expected.txt

Test that inherit and initial works on unprefixed animations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing inherit.
PASS computedStyle.animationName is 'anim'
PASS computedStyle.webkitAnimationName is 'anim'
PASS computedStyle.animationDuration is '5s'
PASS computedStyle.webkitAnimationDuration is '5s'
PASS computedStyle.animationTimingFunction is 'linear'
PASS computedStyle.webkitAnimationTimingFunction is 'linear'
PASS computedStyle.animationDelay is '2s'
PASS computedStyle.webkitAnimationDelay is '2s'
PASS computedStyle.animationIterationCount is 'infinite'
PASS computedStyle.webkitAnimationIterationCount is 'infinite'
PASS computedStyle.animationDirection is 'alternate'
PASS computedStyle.webkitAnimationDirection is 'alternate'
PASS computedStyle.animationPlayState is 'running'
PASS computedStyle.webkitAnimationPlayState is 'running'
Testing initial.
PASS computedStyle.animationName is 'none'
PASS computedStyle.webkitAnimationName is 'none'
PASS computedStyle.animationDuration is '0s'
PASS computedStyle.webkitAnimationDuration is '0s'
PASS computedStyle.animationTimingFunction is 'ease'
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS computedStyle.animationDelay is '0s'
PASS computedStyle.webkitAnimationDelay is '0s'
PASS computedStyle.animationIterationCount is '1'
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS computedStyle.animationDirection is 'normal'
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS computedStyle.animationPlayState is 'running'
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS successfullyParsed is true

TEST COMPLETE
